prevalidations tests minor refactoring
