+++ phpunit.xml
@@ -11,7 +11,7 @@
   <php>
     <env name="APP_ENV" value="testing"/>
     <env name="BCRYPT_ROUNDS" value="4"/>
-    <env name="CACHE_DRIVER" value="file"/>
+    <env name="CACHE_DRIVER" value="array"/>
     <!-- <env name="DB_CONNECTION" value="sqlite"/> -->
     <!-- <env name="DB_DATABASE" value=":memory:"/> -->
     <env name="MAIL_MAILER" value="array"/>
