+++ tests/TestCase.php
@@ -6,15 +6,16 @@ use App\Models\Implementation;
 use App\Services\MailDatabaseLoggerService\Traits\AssertsSentEmails;
 use App\Traits\DoesTesting;
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
-use Illuminate\Support\Facades\Cache;
+use Tests\Traits\MakesApiRequests;
 use Tests\Traits\MakesTestIdentities;
 
 abstract class TestCase extends BaseTestCase
 {
-    use CreatesApplication;
     use DoesTesting;
-    use MakesTestIdentities;
+    use MakesApiRequests;
     use AssertsSentEmails;
+    use CreatesApplication;
+    use MakesTestIdentities;
 
     /**
      * @var string[]
@@ -31,7 +32,6 @@ abstract class TestCase extends BaseTestCase
     {
         parent::setUp();
         Implementation::clearMemo();
-        Cache::flush();
     }
 
     /**
