+++ tests/Traits/MakesTestFunds.php
@@ -44,36 +44,10 @@ trait MakesTestFunds
         ?string $primaryKey = null,
     ): Prevalidation {
         // create prevalidation
-        $response = $this->makePrevalidationForTestCriteriaRequest($organization, $fund, $primaryKey);
+        $response = $this->apiMakePrevalidationForTestCriteriaRequest($organization, $fund, $primaryKey);
         $response->assertSuccessful();
 
         return Prevalidation::find($response->json('data.id'));
-    }
-
-    /**
-     * @param Organization $organization
-     * @param Fund $fund
-     * @param string|null $primaryKey
-     * @return TestResponse
-     */
-    public function makePrevalidationForTestCriteriaRequest(
-        Organization $organization,
-        Fund $fund,
-        ?string $primaryKey = null,
-    ): TestResponse {
-        return $this->makeStorePrevalidationRequest($organization, $fund, [
-            $this->makeRequestCriterionValue($fund, 'test_bool', 'Ja'),
-            $this->makeRequestCriterionValue($fund, 'test_iban', fake()->iban),
-            $this->makeRequestCriterionValue($fund, 'test_date', '01-01-2010'),
-            $this->makeRequestCriterionValue($fund, 'test_email', fake()->email),
-            $this->makeRequestCriterionValue($fund, 'test_string', 'lorem_ipsum'),
-            $this->makeRequestCriterionValue($fund, 'test_string_any', 'ipsum_lorem'),
-            $this->makeRequestCriterionValue($fund, 'test_number', 7),
-            $this->makeRequestCriterionValue($fund, 'test_select', 'foo'),
-            $this->makeRequestCriterionValue($fund, 'test_select_number', 2),
-        ], [
-            $fund->fund_config->csv_primary_key => $primaryKey ?: token_generator()->generate(32),
-        ]);
     }
 
     /**
@@ -195,54 +169,6 @@ trait MakesTestFunds
     }
 
     /**
-     * @param Organization $organization
-     * @param Fund $fund
-     * @param array $records
-     * @param array $extraData
-     * @return TestResponse
-     */
-    protected function makeStorePrevalidationRequest(
-        Organization $organization,
-        Fund $fund,
-        array $records,
-        array $extraData = [],
-    ): TestResponse {
-        $proxy = $this->makeIdentityProxy($organization->identity);
-        $criteria = $fund->criteria()->pluck('record_type_key', 'id')->toArray();
-
-        return $this->postJson("/api/v1/platform/organizations/$organization->id/prevalidations", [
-            'fund_id' => $fund->id,
-            'data' => [
-                ...array_reduce($records, fn ($list, $record) => [
-                    ...$list,
-                    $criteria[$record['fund_criterion_id']] => $record['value'],
-                ], []),
-                ...$extraData,
-            ],
-        ], $this->makeApiHeaders($proxy));
-    }
-
-    /**
-     * @param Organization $organization
-     * @param Fund $fund
-     * @param array $data
-     * @param array $extraData
-     * @return TestResponse
-     */
-    protected function makeStorePrevalidationBatchRequest(
-        Organization $organization,
-        Fund $fund,
-        array $data,
-        array $overwrite = [],
-    ): TestResponse {
-        return $this->postJson("/api/v1/platform/organizations/$organization->id/prevalidations/collection", [
-            'fund_id' => $fund->id,
-            'data' => $data,
-            'overwrite' => $overwrite,
-        ], $this->makeApiHeaders($this->makeIdentityProxy($organization->identity)));
-    }
-
-    /**
      * @param Fund $fund
      * @return void
      */
@@ -316,13 +242,13 @@ trait MakesTestFunds
         if ($type === $recordType::TYPE_SELECT) {
             $recordType->record_type_options()->firstOrCreate([
                 'value' => 'foo',
-            ])->translateOrNew(app()->getLocale())->fill([
+            ])->translateOrNew(app()->getLocale())->forceFill([
                 'name' => 'Foo',
             ])->save();
 
             $recordType->record_type_options()->firstOrCreate([
                 'value' => 'bar',
-            ])->translateOrNew(app()->getLocale())->fill([
+            ])->translateOrNew(app()->getLocale())->forceFill([
                 'name' => 'Bar',
             ])->save();
         }
@@ -330,13 +256,13 @@ trait MakesTestFunds
         if ($type === $recordType::TYPE_SELECT_NUMBER) {
             $recordType->record_type_options()->firstOrCreate([
                 'value' => 1,
-            ])->translateOrNew(app()->getLocale())->fill([
+            ])->translateOrNew(app()->getLocale())->forceFill([
                 'name' => 'Foo',
             ])->save();
 
             $recordType->record_type_options()->firstOrCreate([
                 'value' => 2,
-            ])->translateOrNew(app()->getLocale())->fill([
+            ])->translateOrNew(app()->getLocale())->forceFill([
                 'name' => 'Bar',
             ])->save();
         }
@@ -398,10 +324,10 @@ trait MakesTestFunds
 
             /** @var FundCriteriaStep $stepModel */
             $stepModel = $stepTitle ?
-                ($fund->criteria_steps()->firstWhere([
+                ($fund->criteria_steps()->where([
                     'title' => $stepTitle,
                     ...$stepFields,
-                ]) ?: $fund->criteria_steps()->forceCreate([
+                ])->first() ?: $fund->criteria_steps()->forceCreate([
                     'title' => $stepTitle,
                     ...$stepFields,
                 ])) : null;
@@ -466,7 +392,7 @@ trait MakesTestFunds
             'scope' => 'webshop',
         ]);
 
-        $tag->translateOrNew(app()->getLocale())->fill([
+        $tag->translateOrNew(app()->getLocale())->forceFill([
             'name' => $tagName,
         ])->save();
